Remove contentblock plugin
Github pages does not support arbitrary custom plugins. This is quite
unfortunate, since I had several places where `content_for` was the
perfect way to solve my problem.

Work around this in kind of a hacky way. Not happy with some of these
fixes, but at least I didn’t have to change the design of the site.

Definitely some things to revisit in the future.